+++ R/geom_edge_diagonal.R
@@ -66,6 +66,9 @@
 #' @param flipped Logical, Has the layout been flipped by reassigning the
 #' mapping of x, y etc?
 #'
+#' @param strength The strength of the curvature of the diagonal. `0` will
+#' result in a straight line while `1` will give the familiar S-shape.
+#'
 #' @author Thomas Lin Pedersen
 #'
 #' @family geom_edge_*
@@ -117,13 +120,13 @@ StatEdgeDiagonal <- ggproto('StatEdgeDiagonal', StatBezier,
   },
   required_aes = c('x', 'y', 'xend', 'yend', 'circular'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'flipped', 'n')
+  extra_params = c('na.rm', 'flipped', 'n', 'strength')
 )
 #' @rdname geom_edge_diagonal
 #'
 #' @export
 geom_edge_diagonal <- function(mapping = NULL, data = get_edges(),
-                               position = 'identity', arrow = NULL,
+                               position = 'identity', arrow = NULL, strength = 1,
                                flipped = FALSE, n = 100, lineend = 'butt',
                                linejoin = 'round', linemitre = 1,
                                label_colour = 'black', label_alpha = 1,
@@ -144,7 +147,7 @@ geom_edge_diagonal <- function(mapping = NULL, data = get_edges(),
       list(
         arrow = arrow, lineend = lineend, linejoin = linejoin,
         linemitre = linemitre, na.rm = FALSE, n = n,
-        interpolate = FALSE, flipped = flipped,
+        interpolate = FALSE, flipped = flipped, strength = strength,
         label_colour = label_colour, label_alpha = label_alpha,
         label_parse = label_parse, check_overlap = check_overlap,
         angle_calc = angle_calc, force_flip = force_flip,
@@ -175,13 +178,13 @@ StatEdgeDiagonal2 <- ggproto('StatEdgeDiagonal2', StatBezier2,
   },
   required_aes = c('x', 'y', 'group', 'circular'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'flipped', 'n')
+  extra_params = c('na.rm', 'flipped', 'n', 'strength')
 )
 #' @rdname geom_edge_diagonal
 #'
 #' @export
 geom_edge_diagonal2 <- function(mapping = NULL, data = get_edges('long'),
-                                position = 'identity', arrow = NULL,
+                                position = 'identity', arrow = NULL, strength = 1,
                                 flipped = FALSE, n = 100, lineend = 'butt',
                                 linejoin = 'round', linemitre = 1,
                                 label_colour = 'black', label_alpha = 1,
@@ -202,7 +205,7 @@ geom_edge_diagonal2 <- function(mapping = NULL, data = get_edges('long'),
       list(
         arrow = arrow, lineend = lineend, linejoin = linejoin,
         linemitre = linemitre, na.rm = FALSE, n = n,
-        interpolate = TRUE, flipped = flipped,
+        interpolate = TRUE, flipped = flipped, strength = strength,
         label_colour = label_colour, label_alpha = label_alpha,
         label_parse = label_parse, check_overlap = check_overlap,
         angle_calc = angle_calc, force_flip = force_flip,
@@ -222,13 +225,13 @@ StatEdgeDiagonal0 <- ggproto('StatEdgeDiagonal0', StatBezier0,
   },
   required_aes = c('x', 'y', 'xend', 'yend', 'circular'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'flipped')
+  extra_params = c('na.rm', 'flipped', 'strength')
 )
 #' @rdname geom_edge_diagonal
 #'
 #' @export
 geom_edge_diagonal0 <- function(mapping = NULL, data = get_edges(),
-                                position = 'identity', arrow = NULL,
+                                position = 'identity', arrow = NULL, strength = 1,
                                 flipped = FALSE, lineend = 'butt',
                                 show.legend = NA, ...) {
   mapping <- complete_edge_aes(mapping)
@@ -242,7 +245,7 @@ geom_edge_diagonal0 <- function(mapping = NULL, data = get_edges(),
     inherit.aes = FALSE,
     params = expand_edge_aes(
       list(
-        arrow = arrow, lineend = lineend, na.rm = FALSE,
+        arrow = arrow, lineend = lineend, na.rm = FALSE, strength = strength,
         flipped = flipped, ...
       )
     )
@@ -261,12 +264,12 @@ create_diagonal <- function(from, to, params) {
     r0 <- sqrt(from$x[from$circular]^2 + from$y[from$circular]^2)
     r1 <- sqrt(to$x[to$circular]^2 + to$y[to$circular]^2)
     root <- r0 == 0 | r1 == 0
-    r_mid <- r0 + (r1 - r0) / 2
+    r_mod <- params$strength * (r1 - r0) / 2
 
-    data2$x[from$circular] <- from$x[from$circular] / (r0 / r_mid)
-    data2$y[from$circular] <- from$y[from$circular] / (r0 / r_mid)
-    data3$x[from$circular] <- to$x[from$circular] / (r1 / r_mid)
-    data3$y[from$circular] <- to$y[from$circular] / (r1 / r_mid)
+    data2$x[from$circular] <- from$x[from$circular] / (r0 / (r0 + r_mod))
+    data2$y[from$circular] <- from$y[from$circular] / (r0 / (r0 + r_mod))
+    data3$x[from$circular] <- to$x[from$circular] / (r1 / (r1 - r_mod))
+    data3$y[from$circular] <- to$y[from$circular] / (r1 / (r1 - r_mod))
 
     data2$x[root] <- from$x[root]
     data2$y[root] <- from$y[root]
@@ -276,12 +279,12 @@ create_diagonal <- function(from, to, params) {
   if (any(!from$circular)) {
     if (params$flipped) {
       h_diff <- from$x[!from$circular] - to$x[!from$circular]
-      data2$x[!from$circular] <- from$x[!from$circular] - h_diff / 2
-      data3$x[!from$circular] <- to$x[!from$circular] + h_diff / 2
+      data2$x[!from$circular] <- from$x[!from$circular] - h_diff / 2 * params$strength
+      data3$x[!from$circular] <- to$x[!from$circular] + h_diff / 2 * params$strength
     } else {
       h_diff <- from$y[!from$circular] - to$y[!from$circular]
-      data2$y[!from$circular] <- from$y[!from$circular] - h_diff / 2
-      data3$y[!from$circular] <- to$y[!from$circular] + h_diff / 2
+      data2$y[!from$circular] <- from$y[!from$circular] - h_diff / 2 * params$strength
+      data3$y[!from$circular] <- to$y[!from$circular] + h_diff / 2 * params$strength
     }
   }
   data <- rbind(from, data2, data3, to)