+++ R/geom_edge_hive.R
@@ -59,10 +59,13 @@
 #' @inheritParams geom_edge_link
 #' @inheritParams ggplot2::geom_path
 #'
-#' @param curvature The curvature of the bezier. Defines the distance from the
+#' @param strength The curvature of the bezier. Defines the distance from the
 #' control points to the midpoint between the start and end node. 1 means the
-#' control points are positioned midway between the nodes, while 0 means it
-#' coincide with the nodes (resulting in straight lines)
+#' control points are positioned halfway between the nodes and the middle of the
+#' two axes, while 0 means it coincide with the nodes (resulting in straight
+#' lines)
+#'
+#' @param curvature Deprecated. Use `strength` instead.
 #'
 #' @author Thomas Lin Pedersen
 #'
@@ -123,20 +126,24 @@ StatEdgeHive <- ggproto('StatEdgeHive', StatBezier,
   },
   required_aes = c('x', 'y', 'xend', 'yend'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'n', 'curvature')
+  extra_params = c('na.rm', 'n', 'strength')
 )
 #' @rdname geom_edge_hive
 #'
 #' @export
 geom_edge_hive <- function(mapping = NULL, data = get_edges(),
                            position = 'identity', arrow = NULL,
-                           curvature = 0.5, n = 100, lineend = 'butt',
+                           strength = 1, n = 100, lineend = 'butt',
                            linejoin = 'round', linemitre = 1,
                            label_colour = 'black', label_alpha = 1,
                            label_parse = FALSE, check_overlap = FALSE,
                            angle_calc = 'rot', force_flip = TRUE,
                            label_dodge = NULL, label_push = NULL,
-                           show.legend = NA, ...) {
+                           show.legend = NA, ..., curvature) {
+  if (!missing(curvature)) {
+    .Deprecated(msg = 'The curvature argument has been deprecated in favour of strength')
+    strength <- curvature * 2
+  }
   mapping <- complete_edge_aes(mapping)
   mapping <- aes_intersect(mapping, aes(x = x, y = y,
                                         xend = xend, yend = yend))
@@ -148,7 +155,7 @@ geom_edge_hive <- function(mapping = NULL, data = get_edges(),
       list(
         arrow = arrow, lineend = lineend, linejoin = linejoin,
         linemitre = linemitre, na.rm = FALSE, n = n,
-        interpolate = FALSE, curvature = curvature,
+        interpolate = FALSE, strength = strength,
         label_colour = label_colour, label_alpha = label_alpha,
         label_parse = label_parse, check_overlap = check_overlap,
         angle_calc = angle_calc, force_flip = force_flip,
@@ -180,20 +187,24 @@ StatEdgeHive2 <- ggproto('StatEdgeHive2', StatBezier2,
   },
   required_aes = c('x', 'y', 'group'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'n', 'curvature')
+  extra_params = c('na.rm', 'n', 'strength')
 )
 #' @rdname geom_edge_hive
 #'
 #' @export
 geom_edge_hive2 <- function(mapping = NULL, data = get_edges('long'),
                             position = 'identity', arrow = NULL,
-                            curvature = 0.5, n = 100, lineend = 'butt',
+                            strength = 1, n = 100, lineend = 'butt',
                             linejoin = 'round', linemitre = 1,
                             label_colour = 'black', label_alpha = 1,
                             label_parse = FALSE, check_overlap = FALSE,
                             angle_calc = 'rot', force_flip = TRUE,
                             label_dodge = NULL, label_push = NULL,
-                            show.legend = NA, ...) {
+                            show.legend = NA, ..., curvature) {
+  if (!missing(curvature)) {
+    .Deprecated(msg = 'The curvature argument has been deprecated in favour of strength')
+    strength <- curvature * 2
+  }
   mapping <- complete_edge_aes(mapping)
   mapping <- aes_intersect(mapping, aes(x = x, y = y,
                                         group = edge.id))
@@ -205,7 +216,7 @@ geom_edge_hive2 <- function(mapping = NULL, data = get_edges('long'),
       list(
         arrow = arrow, lineend = lineend, linejoin = linejoin,
         linemitre = linemitre, na.rm = FALSE, n = n,
-        interpolate = TRUE, curvature = curvature,
+        interpolate = TRUE, strength = strength,
         label_colour = label_colour, label_alpha = label_alpha,
         label_parse = label_parse, check_overlap = check_overlap,
         angle_calc = angle_calc, force_flip = force_flip,
@@ -225,14 +236,18 @@ StatEdgeHive0 <- ggproto('StatEdgeHive0', StatBezier0,
   },
   required_aes = c('x', 'y', 'xend', 'yend'),
   default_aes = aes(filter = TRUE),
-  extra_params = c('na.rm', 'curvature')
+  extra_params = c('na.rm', 'strength')
 )
 #' @rdname geom_edge_hive
 #'
 #' @export
 geom_edge_hive0 <- function(mapping = NULL, data = get_edges(),
-                            position = 'identity', arrow = NULL, curvature = 0.5,
-                            lineend = 'butt', show.legend = NA, ...) {
+                            position = 'identity', arrow = NULL, strength = 1,
+                            lineend = 'butt', show.legend = NA, ..., curvature) {
+  if (!missing(curvature)) {
+    .Deprecated(msg = 'The curvature argument has been deprecated in favour of strength')
+    strength <- curvature * 2
+  }
   mapping <- complete_edge_aes(mapping)
   mapping <- aes_intersect(mapping, aes(x = x, y = y,
                                         xend = xend, yend = yend))
@@ -243,7 +258,7 @@ geom_edge_hive0 <- function(mapping = NULL, data = get_edges(),
     params = expand_edge_aes(
       list(
         arrow = arrow, lineend = lineend, na.rm = FALSE,
-        curvature = curvature, ...
+        strength = strength, ...
       )
     )
   )
@@ -268,8 +283,8 @@ create_hive_bezier <- function(from, to, params) {
   middle_axis2 <- ifelse(from_first, to_axis, from_axis)
   middle_axis2 <- ifelse(middle_axis2 < middle_axis1, middle_axis2 + 2 * pi, middle_axis2)
   mean_axis <- (middle_axis2 - middle_axis1) / 2
-  middle_axis1 <- middle_axis1 + mean_axis * params$curvature
-  middle_axis2 <- middle_axis2 - mean_axis * params$curvature
+  middle_axis1 <- middle_axis1 + mean_axis * params$strength / 2
+  middle_axis2 <- middle_axis2 - mean_axis * params$strength / 2
   node_r2 <- sqrt(data2$x^2 + data2$y^2)
   node_r3 <- sqrt(data3$x^2 + data3$y^2)
   data2$x <- node_r2 * cos(ifelse(from_first, middle_axis1, middle_axis2))